v0.4.0:#13: Adds execution backend option.
Enables the selection of TBB as an execution backend.

This change allows users to leverage Intel's Threading Building Blocks (TBB)
for improved parallel execution capabilities. It introduces a new CMake
option and Conan option to control the selected execution backend.
When TBB is selected, the necessary TBB dependencies are automatically
added.